matching https://app.zeplin.io/project.html#pid=57f69abd267208f65efa3e09&sid=57f69f40c94a5e41575ce2a6
fix rendering of constraints in title
