+++ app/assets/stylesheets/bootstrap-overrides.scss
@@ -1,0 +1,26 @@
+// Links
+
+// --- in another world, an accessibility expert is telling us to underline all links
+// a:not(.btn) {
+//   color: $link-color;
+//   text-decoration: underline;
+
+//   &:hover,
+//   &:focus {
+//     color: $link-hover-color;
+//     text-decoration: $link-hover-decoration;
+//   }
+
+//   &:focus {
+//     @include tab-focus;
+//   }
+// }
+
+.btn {
+  border: none;
+  color: $btn-default-color;
+}
+
+.btn.btn-link {
+  color: $link-color;
+}
