+++ app/assets/stylesheets/fishrappr.scss
@@ -1,0 +1,6 @@
+@import 'modules/buttons';
+@import 'modules/custom';
+@import 'modules/graph';
+@import 'modules/grid';
+@import 'modules/modals';
+@import 'modules/viewer';
