+++ app/assets/stylesheets/modules/buttons.scss
@@ -1,8 +1,9 @@
-// @import "globals";
-
 .btn-action {
     background: #4E4E4E;
     color: #EEE;
+    border: none;
+
+    box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.48);
 
     &:hover {
         background: #353535;
@@ -12,16 +13,33 @@
 
 .btn-search-primary {
     background: #E5E5E5;
-    border: 1px solid #B8B8B8;
-    color: #404040;
+
+    border-radius: 3px;
+    box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.36);
 
     &:hover {
         color: #272727;
-        background: #ccc; 
-        border: 1px solid #272727;
+        background: #ccc;
+        box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.50);
+    }
+}
+
+.btn-page-navigation {
+    border-radius: 30px;
+    border: solid 2px #126dc1;
+    min-width: 8em;
+
+    span {
+        color: #0c5292;
     }
 }
 
 .btn-pagination {
     background: transparent;
+}
+
+@media (max-width: $screen-xs-max) {
+    .btn-page-navigation {
+        min-width: 0;
+    }
 }
