+++ app/assets/stylesheets/modules/custom.scss
@@ -532,6 +532,7 @@ div.top-panel-heading {
   display:none;
   border: 3px solid #337ab7;
   z-index: 100;
+  color: $gray-lighter;
 }
 
 .back-to-top-text {
