+++ app/assets/stylesheets/modules/viewer.scss
@@ -20,18 +20,18 @@
 
 .image-viewer-toolbar {
     position: absolute;
-    top: 5px;
+    top: 0; // 5px;
     left: 0;
     right: 0;
-    padding-top: 5px;
-    padding-bottom: 5px;
-    z-index: 9000;
+    padding-top: 14px; // 5px;
+    padding-bottom: 14px; // 5px;
+    z-index: 5000;
     background: $image-viewer-bgcolor;
 
     div > span {
         color: white;
         display: inline-block;
-        padding: 5px 10px;
+        padding: 5px 18px;
         line-height: 18px;
 
         a {
@@ -41,20 +41,27 @@
     }
 
     label {
-        font-size: 12px;
+        font-size: 16px;
         color: white;
         font-weight: normal;
         margin-right: 10px;
-    }
 
-    .btn {
-        // background: #9B9B9B;
-        // color: white;
-        min-width: 100px;
-        font-size: 12px;
+        padding: 5px 0px;
+        line-height: 18px;
+
+        input {
+            margin-right: 4px;
+        }
     }
 
     .btn-action {
+
+        border-radius: 3px;
+        background-color: #6e6e6e;
+        box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.36);
+
+        min-width: 8em;
+
         &:hover {
             background: #3d3d3d;
         }
@@ -62,41 +69,38 @@
 }
 
 .zoom-toolbar > span {
-    border: none; // 1px solid white;
-    font-size: 12px;
+    border: none; 
 }
 
-.navigation-toolbar .pager {
-    margin: 0;
+.blacklight-catalog-show {
+    .navigation-toolbar {
+        padding-top: 18px;
+        padding-bottom: 18px;
 
-    li {
-
-        > span {
-            background: transparent;
-            border: 1px solid #666;
-            min-width: 140px;
+        .btn-issue-link {
+            padding-right: 2px;
+            padding-left: 16px;
         }
 
-        a {
-            background: transparent;
-            border: 1px solid #136DC1;
-            color: #136DC1;
-            min-width: 140px;
-
-            &:hover {
-                background: rgba(19, 109, 193, 0.2);
-            }
+        .row > div > span {
+            display: inline-block;
+            margin: 0;
+            padding: 9px 2px;
+            vertical-align: middle;
         }
 
-        span.current-page-number {
-            background: transparent;
-            border: none;
-            min-width: 0;
-        }        
+        .current-page-number {
+            padding-left: 8px !important;
+            padding-right: 8px !important;
+        }
+
+        .navigate-pages {
+            text-align: right;
+        }
+
     }
-
-
 }
+
 
 .local-options-toolbar {
     display: flex;
@@ -138,6 +142,7 @@
         h3 {
             font-size: 14px;
             font-weight: normal;
+            margin-top: 0;
         }
 
         .panel-heading {
@@ -147,6 +152,14 @@
 
         .panel-body {
             padding-top: 8px;
+        }
+
+        margin-bottom: -5px;
+
+        &:nth-of-type(1) {
+            h3 {
+                margin-top: $line-height-computed;
+            }
         }
     }
 
@@ -189,8 +202,12 @@
 }
 
 .panel-actions-page .btn {
-    background: #136DC1;
+    // background: #136DC1;
     color: white;
+
+    border-radius: 3px;
+    background-color: #126dc1;
+    // box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.36);
 
     &:hover {
         background: #0e5393;
@@ -198,8 +215,10 @@
 }
 
 .panel-actions-issue .btn {
-    background: #0C5292;
     color: white;
+
+    background-color: #0c5292;
+    // box-shadow: inset 0 -2px 0 0 rgba(0, 0, 0, 0.48);
 
     &:hover {
         background: #083863;
@@ -223,15 +242,33 @@
     border-radius: 4px;
 }
 
+.alert-plaintext-disclaimer {
+    border-radius: 3px;
+    background-color: #f2f2f2;
+    border: solid 2px #6e6e6e;
+    color: #4e4e4e;
+    padding: $alert-padding;
+    margin-bottom: $line-height-computed;
+
+    a {
+        color: #4e4e4e;
+        text-decoration: underline;
+
+        &:hover {
+            // font-weight: bold;
+            color: black;
+        }
+    }
+}
+
 @media (max-width: $screen-xs-max) {
-    .image-viewer-toolbar .btn {
-        min-width: 1em;
+
+    #image-viewer {
+        height: 95%;
     }
 
-    .navigation-toolbar .pager li {
-        a, span {
-            min-width: 1em;
-        }
+    .image-viewer-toolbar .btn {
+        min-width: 1em;
     }
 
     .btn-issue-link {
@@ -241,4 +278,13 @@
     .navbar-brand img {
         max-width: 100%;
     }
+
+    .blacklight-catalog-show {
+        .navigation-toolbar {
+            .navigate-pages {
+                text-align: center;
+            }
+        }        
+    }
+
 }
