+++ app/helpers/zzzzzz_helper.rb
@@ -21,26 +21,5 @@ module ZzzzzzHelper
     end
   end
 
-  def render_search_to_page_title(params)
-    constraints = []
 
-    if params['q'].present?
-      q_label = label_for_search_field(params[:search_field]) unless default_search_field && params[:search_field] == default_search_field[:key]
-
-      constraints += if q_label.present?
-                       [t('blacklight.search.page_title.constraint', label: q_label, value: params['q'])]
-                     else
-                       [params['q']]
-                     end
-    end
-
-    if params['f'].present?
-      constraints += params['f'].to_unsafe_h.collect { |key, value| render_search_to_page_title_filter(key, value) }
-    end
-
-    retval = constraints.join(' / ')
-    unless retval.blank?
-      retval += " - "
-    end
-  end
 end
