+++ app/views/catalog/_show_default.html.erb
@@ -1,23 +1,6 @@
-<%# default partial to display solr document fields in catalog show view -%>
+<div class="alert-plaintext-disclaimer">
+  <%= raw(t('blacklight.show.disclaimer')) %>
+</div>
 <article>
   <%== render_plain_text(@document, 'page_text') %>
 </article>
-
-<% if false %>
-<dl class="dl-horizontal  dl-invert">
-  <% document_show_fields(document).each do |field_name, field| -%>
-    <% if should_render_show_field? document, field %>
-        <% if field.key == "page_text" %>
-          <dt class="blacklight">Page Text: </dt> 
-          <% if false %>
-          <dd class="blacklight"><%== highlight_search_terms(@document['page_text']) %> </dd>     
-          <% end %>
-          <dd class="blacklight"><%== render_plain_text(@document, 'page_text') %></dd>
-         <% else %>
-	       <dt class="blacklight-<%= field_name.parameterize %>"><%= render_document_show_field_label document, field: field_name %></dt>
-	        <dd class="blacklight-<%= field_name.parameterize %>"><%= render_document_show_field_value document, field: field_name %></dd>
-	     <% end %>   
-    <% end -%>
-  <% end -%>
-</dl>
-<% end %>
